==> src/dspg17/ckelling/ScrapingCode/depsy/02_error_detection.Rmd <==
---
title: "Depsy Scraping"
author: "Claire Kelling"
date: "6/26/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

First, I will load the list of all packages that I have scraped from the Cran list. 
```{r load nodes}
library(rvest)
library(rjson)
load(file= '~/git/oss/data/oss/original/depsy/all_packages_cran.Rdata')
```

In this file, I will create a list of errors- where the package is not listed on Depsy, but it is on cran.
```{r nodal features}
#API for GGplot: http://depsy.org/api/package/cran/ggplot2
name_vec <- as.vector(all_packages$oss_name)

test <- c()
error_vec <- c()

error_message<- "Error in"
for(i in 1:length(name_vec)){
  link <- paste('http://depsy.org/api/package/cran/', name_vec[i], sep='')
  print(i)
  test <- try(read_html(link))
  error_test <- substr(test[1], 1, 8)
  if(error_test == error_message){
    error_vec <- rbind(error_vec, i)
  }
}

length(error_vec)
save(error_vec, file = "~/git/oss/data/oss/original/depsy/error_vector.Rdata")
```
